--- metrics/IMeasurement.hpp
@@ -1,74 +1,0 @@
-/*##############################################################################
-
-    HPCC SYSTEMS software Copyright (C) 2020 HPCC Systems®.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-############################################################################## */
-
-#pragma once
-
-#include <string>
-#include "../../system/jlib/jiface.hpp"
-
-namespace hpccMetrics
-{
-
-//
-// Defines the type of the measurement value string returned by valueToString
-enum ValueType {
-    NONE,
-    STRING,
-    LONG,
-    INTEGER,
-    DOUBLE,
-    FLOAT,
-    DATE,
-    DATE_NANOS,
-    BOOLEAN
-} ;
-
-
-enum MetricType {
-    CUSTOM,
-    COUNTER,
-    RATE,
-    GAUGE
-};
-
-//
-// Interface for object used to hold and report a metric value obtained during collection
-interface IMeasurement
-{
-
-    //
-    // Return the value as a string object
-    virtual std::string valueToString() const = 0;
-
-    //
-    // Return the name of the metric value for reporting purposes
-    virtual const std::string &getName() const = 0;
-
-    //
-    // Return the metric type
-    virtual MetricType getMetricType() const = 0;
-
-    //
-    // Return the metric description
-    virtual const std::string &getDescription() const = 0;
-
-    //
-    // Get the metric value type
-    virtual ValueType getValueType() const = 0;
-};
-
-}
