+++ metrics/sinks/CMakeLists.txt
@@ -14,6 +14,4 @@
 #    limitations under the License.
 ################################################################################
 
-#HPCC_ADD_SUBDIRECTORY (ELK)
 HPCC_ADD_SUBDIRECTORY (file)
-HPCC_ADD_SUBDIRECTORY (elasticsearch)
